style(src): split move sandboxer module into another file
